requirements.txt
argparse
face_recognition
kafka_python